updated & rebuilt lib with unified client
